suzerain: #2729 add cast for icc 12.1 on cab
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@42336 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ tests/test_bl.cpp
@@ -252,7 +252,7 @@ const double ChallengingFixture::coeffs_H0[ChallengingFixture::ndof] = {
 BOOST_FIXTURE_TEST_CASE( find_edge, ChallengingFixture )
 {
     // Sanity check the provided data amount
-    BOOST_REQUIRE_EQUAL(b.n(), ndof);
+    BOOST_REQUIRE_EQUAL(b.n(), static_cast<int>(ndof));
 
     // Prepare working storage
     shared_ptr<gsl_matrix> dB(gsl_matrix_alloc(b.k(), 3), gsl_matrix_free);
